move ftml-components util functions to wikijump/util
