+++ web/modules/ftml-components/src/components/math/math.ts
@@ -1,5 +1,5 @@
+import { addElement, observe, pauseObservation } from "@wikijump/util"
 import { BaseTooltipButton } from "../../base-tooltip-button"
-import { addElement, observe, pauseObservation } from "../../util"
 
 const NEED_TO_POLYFILL = !hasMathMLSupport()
